Move constant to summary classes
