+++ src/main/java/com/amazon/opendistro/elasticsearch/performanceanalyzer/rca/framework/util/RcaConsts.java
@@ -77,9 +77,4 @@ public class RcaConsts {
     public static final String LOCUS_DATA_NODE = "data-node";
     public static final String LOCUS_MASTER_NODE = "master-node";
   }
-  
-  public static final String HOT_NODE_SUMMARY_TABLE = HotNodeSummary.class.getSimpleName();
-  public static final String HOT_CLUSTER_SUMMARY_TABLE = HotClusterSummary.class.getSimpleName();
-  public static final String HOT_RESOURCE_SUMMARY_TABLE = HotResourceSummary.class.getSimpleName();
-  public static final String HIGH_HEAP_USAGE_CLUSTER_RCA_TABLE = HighHeapUsageClusterRca.class.getSimpleName();
 }
